+++ homeassistant/components/tuya/switch.py
@@ -27,7 +27,9 @@ from homeassistant.helpers.issue_registry import (
 from . import TuyaConfigEntry
 from .const import DOMAIN, TUYA_DISCOVERY_NEW, DeviceCategory, DPCode
 from .entity import TuyaEntity
-from .tuya_device_handlers import TUYA_QUIRKS_REGISTRY, TuyaSwitchDefinition, parse_enum
+from .tuya_device_handlers import TUYA_QUIRKS_REGISTRY
+from .tuya_device_handlers.builder import TuyaSwitchDefinition
+from .tuya_device_handlers.helpers import parse_enum
 
 
 @dataclass(frozen=True, kw_only=True)
