+++ homeassistant/components/tuya/tuya_device_handlers/builder/__init__.py
@@ -1,0 +1,23 @@
+"""Tuya device handler."""
+
+from __future__ import annotations
+
+from .base_quirk import (
+    TuyaClimateDefinition,
+    TuyaCoverDefinition,
+    TuyaDeviceQuirk,
+    TuyaIntegerConversionFunction,
+    TuyaSelectDefinition,
+    TuyaSensorDefinition,
+    TuyaSwitchDefinition,
+)
+
+__all__ = [
+    "TuyaClimateDefinition",
+    "TuyaCoverDefinition",
+    "TuyaDeviceQuirk",
+    "TuyaIntegerConversionFunction",
+    "TuyaSelectDefinition",
+    "TuyaSensorDefinition",
+    "TuyaSwitchDefinition",
+]
