+++ homeassistant/components/tuya/tuya_device_handlers/helpers/const.py
