+++ homeassistant/components/tuya/tuya_device_handlers/helpers/homeassistant.py
