+++ homeassistant/components/tuya/tuya_device_handlers/helpers/models.py
@@ -1,0 +1,41 @@
+"""Tuya models."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+
+@dataclass
+class TuyaIntegerTypeDefinition:
+    """Definition of an integer type data."""
+
+    dpcode: str
+    min: int
+    max: int
+    scale: float
+    step: float
+    unit: str | None = None
+    type: str | None = None
+
+    @property
+    def max_scaled(self) -> float:
+        """Return the max scaled."""
+        return self.scale_value(self.max)
+
+    @property
+    def min_scaled(self) -> float:
+        """Return the min scaled."""
+        return self.scale_value(self.min)
+
+    @property
+    def step_scaled(self) -> float:
+        """Return the step scaled."""
+        return self.step / (10**self.scale)
+
+    def scale_value(self, value: float) -> float:
+        """Scale a value."""
+        return value / (10**self.scale)
+
+    def scale_value_back(self, value: float) -> int:
+        """Return raw value for scaled."""
+        return int(value * (10**self.scale))
