+++ homeassistant/components/tuya/tuya_device_handlers/helpers/utils.py
